requirements.txt
paho-mqtt
pyYAML
opencv-python
numpy
pyserial